Rename controller => addressController and add event click on label focus on input
