+++ address-book-client/src/sections/modal/index.html
@@ -4,27 +4,27 @@
   </div>
   <div class="modal__inputs">
     <div class="modal__inputs__field">
-      <label class="text text--gray text--bold text--xl">Name *</label>
+      <label for="name" class="text text--gray text--bold text--xl">Name *</label>
       <input id="name" type="text" name="name" class="input text text--normal text--lg" />
       <span class="warning-text text text--warning text--xs text--medium"></span>
     </div>
     <div class="modal__inputs__field">
-      <label class="text text--gray text--bold text--xl">Relation *</label>
+      <label for="relation" class="text text--gray text--bold text--xl">Relation *</label>
       <input id="relation" type="text" name="relation" class="input text text--normal text--lg" />
       <span class="warning-text text text--warning text--xs text--medium"></span>
     </div>
     <div class="modal__inputs__field">
-      <label class="text text--gray text--bold text--xl">Phone number *</label>
+      <label for="phone" class="text text--gray text--bold text--xl">Phone number *</label>
       <input id="phone" type="number" name="phone" class="text text--normal text--lg" />
       <span class="warning-text text text--warning text--xs text--medium"></span>
     </div>
     <div class="modal__inputs__field">
-      <label class="text text--gray text--bold text--xl">Image *</label>
+      <label for="avatar" class="text text--gray text--bold text--xl">Image *</label>
       <input id="avatar" name="avatar" class="text text--normal text--lg" />
       <span class="warning-text text text--warning text--xs text--medium"></span>
     </div>
     <div class="modal__inputs__field">
-      <label class="text text--gray text--bold text--xl">Email *</label>
+      <label for="email" class="text text--gray text--bold text--xl">Email *</label>
       <input id="email" type="text" name="email" class="text text--normal text--lg" />
       <span class="warning-text text text--warning text--xs text--medium"></span>
     </div>
